% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algalViz.R
\name{algalViz}
\alias{algalViz}
\title{algalViz}
\usage{
algalViz(data, proportion = T, count_column = "Count", facet_var = NULL)
}
\arguments{
\item{data}{A dataframe}

\item{proportion}{boolean, wether to report proportions or total counts}

\item{count_column}{string, column name of count data}
}
\value{
ggplot ggplot vizualization object
}
\description{
utility function to help in vizualizing algal distributions across sites
}
\examples{
algalViz(data= df, standardize = T, count_value = "Count")
}
