% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrientViz.R
\name{cleanNuts}
\alias{cleanNuts}
\title{Nutrient data workup}
\usage{
cleanNuts(data)
}
\arguments{
\item{data}{}
}
\value{
tall clean dataframe for vizualization
}
\description{
Preprocessing of nutrient data for visualization. Specifically, it removes blanks, drop nas,
and blanks, specifies datatypes, renames variables, and finds the mean for each site, date,
parameter combination. Requires columns, Type, Date, Site, Round, Year in any order
}
\examples{
cleanNuts(data)
}
