% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrientViz.R
\name{nutsThroughTime}
\alias{nutsThroughTime}
\title{nutsThroughTime}
\usage{
nutsThroughTime(data)
}
\arguments{
\item{dataframe}{with site, date, and parameters in other columns}
}
\value{
ggplot object
}
\description{
Creates a ggplot object of nutrient data through time. Specifically, it creates a ggplot object
}
\examples{
nutsThroughTime(data)
nutsThroughTime(df)
}
