% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterColumnViz.R
\name{waterColumnViz}
\alias{waterColumnViz}
\title{waterColumnViz}
\usage{
waterColumnViz(df, variable)
}
\arguments{
\item{df}{dataframe with sites, rounds, depth, and water quality variables
of interes in long format}

\item{variable}{string specifyning which variable to vizualize}
}
\value{
ggplot plot
}
\description{
Vizualize water quality variables along the water column for multiple time
points and multiple site
}
\examples{
myfunction(X,Y)
}
