% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterColumnViz.R
\name{waterColumnViz}
\alias{waterColumnViz}
\title{waterColumnViz}
\usage{
waterColumnViz(df, parameter, xlabel)
}
\arguments{
\item{df}{dataframe with round, date, site, depth, followed by parameter columns}

\item{variable}{string specifyning which variable to vizualize}
}
\value{
ggplot plot
}
\description{
Vizualize water quality variables along the water column for multiple time
points and multiple site
}
\details{
A function that helps water quality data vizualization for a single water quality
parameter measured over multiple sites and depths and days
}
\examples{
myfunction(X,Y)
}
